fix allure-js-commons tests
